back/requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==3.2.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
